Add NA object
NA still has a few bugs (namely, await and equality still don't work)
but it's mostly functioning.

The intention with Walter is to allow a config block to continue
execution through to the end despite errors, so that all errors can be
collected and reported rather than just the first one; NA is going to be
a critical part of allowing that to work.